split one schema file into several
