+++ backend/src/mood_tracker/presentation/api/auth/routers/refresh.py
@@ -4,7 +4,9 @@ from fastapi import APIRouter, Cookie, HTTPException, Response, status
 from mood_tracker.application.auth.dto.refresh_user import RefreshUserInputDTO
 from mood_tracker.application.auth.use_cases import RefreshUserUseCase
 from mood_tracker.constants import REFRESH_TOKEN_COOKIE_NAME
-from mood_tracker.presentation.api.auth.schemas.auth import UserRefreshResponse
+from mood_tracker.presentation.api.auth.schemas.refresh import (
+    UserRefreshResponse,
+)
 from mood_tracker.presentation.api.cookie_service import CookieService
 
 router = APIRouter()
