+++ backend/src/mood_tracker/presentation/api/auth/schemas/login.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel, EmailStr
+
+
+class UserLoginRequest(BaseModel):
+    email: EmailStr
+    password: str
+
+
+class UserLoginResponse(BaseModel):
+    access_token: str
