+++ backend/src/mood_tracker/presentation/api/auth/schemas/refresh.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class UserRefreshResponse(BaseModel):
+    access_token: str
