+++ backend/src/mood_tracker/presentation/api/auth/schemas/register.py
@@ -34,16 +34,3 @@ class UserRegisterRequest(BaseModel):
 
 class UserRegisterResponse(BaseModel):
     access_token: str
-
-
-class UserLoginRequest(BaseModel):
-    email: EmailStr
-    password: str
-
-
-class UserLoginResponse(BaseModel):
-    access_token: str
-
-
-class UserRefreshResponse(BaseModel):
-    access_token: str
